sample_project/requirements.txt
pandas==2.2.3
numpy==2.1.3
python-dateutil==2.9.0.post0